refactor(backend): extract treat untreated tracks service from tracks service
